+++ myceliumd-common/src/lib.rs
@@ -244,12 +244,6 @@ pub enum RoutesCommand {
         #[arg(long = "json", default_value_t = false)]
         json: bool,
     },
-    /// Print all subnets which are explicitly marked as not having a route
-    NoRoute {
-        /// Print subnets in JSON format
-        #[arg(long = "json", default_value_t = false)]
-        json: bool,
-    },
 }
 
 #[derive(Debug, Subcommand)]
@@ -795,9 +789,6 @@ pub async fn dispatch_subcommand(
             }
             RoutesCommand::Queried { json } => {
                 return mycelium_cli::list_queried_subnets(api_addr, json).await;
-            }
-            RoutesCommand::NoRoute { json } => {
-                return mycelium_cli::list_no_route_entries(api_addr, json).await;
             }
         },
         Command::Proxy { command } => match command {